block-weight added

+++ Yellow/average-temperature-2/main.cpp
@@ -4,11 +4,11 @@
 using namespace std;
 
 int main() {
-    vector<int> vec;
-    vector<int> vecGavg;
-    int n = 0;
-    int temp = 0;
-    int sum = 0;
+    vector<long> vec;
+    vector<long> vecGavg;
+    long n = 0;
+    long temp = 0;
+    long sum = 0;
     cin >> n;
 
     while(n-- > 0) {
@@ -17,7 +17,7 @@ int main() {
         sum += temp;
     }
 
-    int avg = sum / static_cast<int>(vec.size());
+    long avg = sum / static_cast<long>(vec.size());
 
     for(size_t i = 0; i < vec.size(); ++i) {
         if(vec[i] > avg) {
